# src/jobfeed/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationRow": {
        "description": "One applications-history row: audit metadata + snapshot hashes only.",
        "properties": {
          "application_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Method"
          },
          "applied_at": {
            "format": "date-time",
            "title": "Applied At",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "master_resume_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Master Resume Hash"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tailored_resume_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailored Resume Hash"
          }
        },
        "required": [
          "job_id",
          "applied_at",
          "application_method",
          "notes",
          "master_resume_hash",
          "tailored_resume_hash"
        ],
        "title": "ApplicationRow",
        "type": "object"
      },
      "ApplicationsListResponse": {
        "description": "``GET /api/applications`` response.",
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/ApplicationRow"
            },
            "title": "Applications",
            "type": "array"
          }
        },
        "required": [
          "applications"
        ],
        "title": "ApplicationsListResponse",
        "type": "object"
      },
      "ApplyResponse": {
        "description": "``POST /api/jobs/{id}/apply`` response.\n\n``applied`` is False for the no-op parity case (already applied);\n``reapply_notice`` is set only on a new apply with a same-company\nactive application.",
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "reapply_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reapply Notice"
          }
        },
        "required": [
          "applied",
          "reapply_notice"
        ],
        "title": "ApplyResponse",
        "type": "object"
      },
...
      "Body_apply_to_job_api_jobs__job_id__apply_post": {
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "cover_letter": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Letter"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "source_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Id"
          },
          "tailored": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailored"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          }
        },
        "title": "Body_apply_to_job_api_jobs__job_id__apply_post",
        "type": "object"
      },
      "Body_apply_to_real_job_api_real_jobs__real_job_id__apply_post": {
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "cover_letter": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Letter"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "source_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Id"
          },
          "tailored": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailored"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          }
        },
        "title": "Body_apply_to_real_job_api_real_jobs__real_job_id__apply_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RealJobApplicationRow": {
        "description": "Canonical submission event with source and target provenance.",
        "properties": {
          "application_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Method"
          },
          "applied_at": {
            "format": "date-time",
            "title": "Applied At",
            "type": "string"
          },
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "real_job_id": {
            "title": "Real Job Id",
            "type": "string"
          },
          "source_applied_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Applied Job Id"
          },
          "source_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Id"
          }
        },
        "required": [
          "id",
          "real_job_id",
          "source_job_id",
          "source_applied_job_id",
          "apply_url",
          "applied_at",
          "application_method",
          "notes"
        ],
        "title": "RealJobApplicationRow",
        "type": "object"
      },
      "RealJobApplicationsListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/RealJobApplicationRow"
            },
            "title": "Applications",
            "type": "array"
          }
        },
        "required": [
          "applications"
        ],
        "title": "RealJobApplicationsListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Jobfeed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/applications": {
      "get": {
        "description": "List recent application records, newest first.\n\nArgs:\n    service: Shared application service from the app state.\n    limit: Maximum number of records to return.\n\nReturns:\n    Application history rows (audit metadata + snapshot hashes only).",
        "operationId": "list_applications_api_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Applications"
      }
    },
...
    "/api/jobs/{job_id}/apply": {
      "post": {
        "description": "Record an application with resume snapshots and audit fields.\n\nMirrors the CLI apply command: master resume read from settings,\nStage B snapshot fields captured via the shared service helper, and the\nno-op parity result when the job was already applied.\n\nArgs:\n    job_id: Store-assigned job identity.\n    request: Current request (carries the app settings).\n    form: Parsed multipart upload + form fields.\n    service: Shared application service from the app state.\n    store: Shared job store (existence check only).\n\nReturns:\n    ``applied`` flag plus the reapply notice when one applies.\n\nRaises:\n    ApiError: 404 unknown job; 422 non-UTF-8 upload; 409 with code\n        ``illegal_transition`` when the job is in a terminal status.",
        "operationId": "apply_to_job_api_jobs__job_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_apply_to_job_api_jobs__job_id__apply_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply To Job"
      }
    },
...
    "/api/real-jobs/applications": {
      "get": {
        "description": "List canonical submission events, including their source provenance.",
        "operationId": "list_real_job_applications_api_real_jobs_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealJobApplicationsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Real Job Applications"
      }
    },
...
    "/api/real-jobs/{real_job_id}/apply": {
      "post": {
        "description": "Record a submission against a canonical job with explicit source.",
        "operationId": "apply_to_real_job_api_real_jobs__real_job_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "real_job_id",
            "required": true,
            "schema": {
              "title": "Real Job Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_apply_to_real_job_api_real_jobs__real_job_id__apply_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply To Real Job"
      }
    },
...
  }
}